refactor(prelude): move log sampling to its own module, use macros
